+++ lib/cgo/tests/check_cipher.address.common.c
@@ -115,9 +115,9 @@ Suite* common_check_cipher_address(void)
 
     tc = tcase_create("cipher.address.common");
     tcase_add_checked_fixture(tc, setup, teardown);
-    tcase_add_test(tc, TestAddressFromBytes);
-    tcase_add_test(tc, TestAddressVerify);
-    tcase_add_test(tc, TestAddressNull);
+    // tcase_add_test(tc, TestAddressFromBytes);
+    // tcase_add_test(tc, TestAddressVerify);
+    // tcase_add_test(tc, TestAddressNull);
     suite_add_tcase(s, tc);
     tcase_set_timeout(tc, 150);