+++ CHANGELOG.md
@@ -86,6 +86,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added function `SKY_bip32_PrivateKey_GetKey`
 - Added function `SKY_bip32_PublicKey_GetKey`
 - Added datatype `coin__HashPair`
+- Added function `SKY_bip32_PrivateKey_GetDepth`
+- Added function `SKY_bip32_PublicKey_GetDepth`
+- Added function `SKY_bip32_PrivateKey_ChildNumber`
+- Added function `SKY_bip32_PublicKey_ChildNumber`
 
 
 ### Removed